+++ src/TestableHttpClient/Utils/TimeoutHttpResponseMessage.cs
@@ -1,0 +1,8 @@
+﻿using System.Net.Http;
+
+namespace TestableHttpClient.Utils
+{
+    internal class TimeoutHttpResponseMessage : HttpResponseMessage
+    {
+    }
+}